plotting format gui update
